LayoutTests/http/tests/messaging/broadcastchannel-partitioning-with-storage-access-expected.txt

Tests that BroadcastChannel's origin partitioning can be relaxed via requestStorageAccess()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Received message from same-origin iframe with storage access under cross-origin popup
PASS gotMessageFromSameOriginIframeUnderCrossOriginPopup is false
PASS successfullyParsed is true

TEST COMPLETE
